refactor: race -> carRace로 변경
